+++ R/input-date.R
@@ -33,6 +33,13 @@
 #'   Can be "month" (the default), "year", or "decade".
 #' @param weekstart Which day is the start of the week. Should be an integer
 #'   from 0 (Sunday) to 6 (Saturday).
+#' @param minviewmode The minimum resolution available when selecting dates.
+#'   Can be "days" (the default), "months", "years", "decades", or "centuries".
+#'   This can restrict date selection to a month, a year, a decade, or a century.
+#'   The selected date will correspond to the first day of the selected period.
+#' @param maxviewmode The maximum resolution available when selecting dates.
+#'   Can be "days", "months", "years", "decades", or "centuries" (the default).
+#'   This can restrict the ability to zoom out for quicker navigation.
 #' @param language The language used for month and day names. Default is "en".
 #'   Other valid values include "ar", "az", "bg", "bs", "ca", "cs", "cy", "da",
 #'   "de", "el", "en-AU", "en-GB", "eo", "es", "et", "eu", "fa", "fi", "fo",
@@ -81,7 +88,10 @@
 #'
 #'   # Disable specific dates.
 #'   dateInput("date8", "Date:", value = "2012-02-29",
-#'             datesdisabled = c("2012-03-01", "2012-03-02"))
+#'             datesdisabled = c("2012-03-01", "2012-03-02")),
+#'
+#'   # Restrict user to selecting a month
+#'   dateInput("date9", "Date:", minviewmode = "months")
 #' )
 #'
 #' shinyApp(ui, server = function(input, output) { })
@@ -93,6 +103,7 @@
 #' @export
 dateInput <- function(inputId, label, value = NULL, min = NULL, max = NULL,
   format = "yyyy-mm-dd", startview = "month", weekstart = 0,
+  minviewmode = "days", maxviewmode = "centuries",
   language = "en", width = NULL, autoclose = TRUE,
   datesdisabled = NULL, daysofweekdisabled = NULL) {
 
@@ -100,6 +111,11 @@ dateInput <- function(inputId, label, value = NULL, min = NULL, max = NULL,
   min <- dateYMD(min, "min")
   max <- dateYMD(max, "max")
   datesdisabled <- dateYMD(datesdisabled, "datesdisabled")
+
+  # datepicker provides aliases that might cause confusion (e.g. "month" => "days")
+  viewmode_choices <- c("days", "months", "years", "decades", "centuries")
+  minviewmode <- rlang::arg_match(minviewmode, viewmode_choices)
+  maxviewmode <- rlang::arg_match(maxviewmode, viewmode_choices)
 
   value <- restoreInput(id = inputId, default = value)
 
@@ -118,6 +134,8 @@ dateInput <- function(inputId, label, value = NULL, min = NULL, max = NULL,
                `data-date-week-start` = weekstart,
                `data-date-format` = format,
                `data-date-start-view` = startview,
+               `data-date-min-view-mode` = minviewmode,
+               `data-date-max-view-mode` = maxviewmode,
                `data-min-date` = min,
                `data-max-date` = max,
                `data-initial-date` = value,