+++ NEWS.md
@@ -203,6 +203,9 @@ shiny 1.5.0
 
 * Fixed #2856: Bumped jQuery 3 from 3.4.1 to 3.5.1. (#2857)
 
+* Resolved [#1365](https://github.com/rstudio/shiny/issues/1365): Added `minviewmode` and `maxviewmode` as new parameters to `dateInput()` and `dateRangeInput()`. This exposes the underlying arguments of [Bootstrap Datepicker](https://bootstrap-datepicker.readthedocs.io/en/latest/options.html#maxviewmode). The `minviewmode` parameter makes it possible to restrict date selection to a month, a year, a decade, or a century.
+
+
 ### Bug fixes
 
 * Fixed #2606: `debounce()` would not work properly if the code in the reactive expression threw an error on the first run. (#2652)